refactor: Use full qualified name for PanicInfo in panic handler
Removed the explicit import of `core::panic::PanicInfo` and updated the panic function to use the fully qualified type path `&core::panic::PanicInfo` instead of `&PanicInfo`. This reduces unnecessary imports and maintains clarity in the type references.
