+++ fullerene-kernel/src/vga/tests.rs
@@ -1,135 +1,12 @@
-// fullerene-kernel/src/vga.rs
-
-use spin::{Mutex, Once};
-use x86_64::instructions::port::Port;
-
-pub use petroleum::{Color, ColorCode, ScreenChar, TextBufferOperations};
-
-const BUFFER_HEIGHT: usize = 25;
-const BUFFER_WIDTH: usize = 80;
-
-/// Represents the VGA text buffer writer.
-pub struct VgaBuffer {
-    buffer: &'static mut [[ScreenChar; BUFFER_WIDTH]; BUFFER_HEIGHT],
-    column_position: usize,
-    row_position: usize,
-    color_code: ColorCode,
-}
-
-impl VgaBuffer {
-    /// Creates a new VgaBuffer instance.
-    pub fn new() -> VgaBuffer {
-        VgaBuffer {
-            buffer: unsafe { &mut *(0xb8000 as *mut _) },
-            column_position: 0,
-            row_position: 0,
-            color_code: ColorCode::new(Color::Green, Color::Black),
-        }
-    }
-
-    /// Updates the hardware cursor position.
-    pub fn update_cursor(&self) {
-        let pos = self.row_position * BUFFER_WIDTH + self.column_position;
-        unsafe {
-            let mut command_port = Port::new(0x3D4);
-            let mut data_port = Port::new(0x3D5);
-
-            command_port.write(0x0F_u8);
-            data_port.write((pos & 0xFF) as u8);
-            command_port.write(0x0E_u8);
-            data_port.write(((pos >> 8) & 0xFF) as u8);
-        }
-    }
-}
-
-impl core::fmt::Write for VgaBuffer {
-    fn write_str(&mut self, s: &str) -> core::fmt::Result {
-        self.write_string(s);
-        Ok(())
-    }
-}
-
-// To ensure thread-safety for `spin` crates.
-unsafe impl Send for VgaBuffer {}
-unsafe impl Sync for VgaBuffer {}
-
-impl TextBufferOperations for VgaBuffer {
-    fn get_width(&self) -> usize {
-        BUFFER_WIDTH
-    }
-
-    fn get_height(&self) -> usize {
-        BUFFER_HEIGHT
-    }
-
-    fn get_color_code(&self) -> ColorCode {
-        self.color_code
-    }
-
-    fn get_position(&self) -> (usize, usize) {
-        (self.row_position, self.column_position)
-    }
-
-    fn set_position(&mut self, row: usize, col: usize) {
-        self.row_position = row;
-        self.column_position = col;
-    }
-
-    fn set_char_at(&mut self, row: usize, col: usize, chr: ScreenChar) {
-        if row < BUFFER_HEIGHT && col < BUFFER_WIDTH {
-            self.buffer[row][col] = chr;
-        }
-    }
-
-    fn get_char_at(&self, row: usize, col: usize) -> ScreenChar {
-        if row < BUFFER_HEIGHT && col < BUFFER_WIDTH {
-            self.buffer[row][col]
-        } else {
-            ScreenChar {
-                ascii_character: 0,
-                color_code: self.color_code,
-            }
-        }
-    }
-
-    fn scroll_up(&mut self) {
-        for row in 1..BUFFER_HEIGHT {
-            for col in 0..BUFFER_WIDTH {
-                self.buffer[row - 1][col] = self.buffer[row][col];
-            }
-        }
-        self.clear_row(BUFFER_HEIGHT - 1);
-    }
-}
-
-// Global singleton for the VGA buffer writer
-pub(crate) static VGA_BUFFER: Once<Mutex<VgaBuffer>> = Once::new();
-
-/// Logs a message to the VGA screen.
-pub fn log(msg: &str) {
-    if let Some(vga) = VGA_BUFFER.get() {
-        let mut writer = vga.lock();
-        writer.write_string(msg);
-        writer.update_cursor();
-    }
-}
-
-/// Initializes the VGA screen.
-pub fn vga_init() {
-    VGA_BUFFER.call_once(|| Mutex::new(VgaBuffer::new()));
-    let mut writer = VGA_BUFFER.get().unwrap().lock();
-    writer.clear_screen(); // Clear screen on boot
-    writer.color_code = ColorCode::new(Color::Green, Color::Black);
-    writer.write_string("Hello QEMU by FullereneOS!\n");
-    writer.color_code = ColorCode::new(Color::Green, Color::Black);
-    writer.write_string("This is output directly to VGA.\n");
-    writer.update_cursor();
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use alloc::{vec, vec::Vec}; // Import the vec! macro and Vec type
+    use alloc::string::String;
+    use alloc::vec;
+    use alloc::vec::Vec; // Import the vec! macro and Vec type
+
+    // Bring in necessary types and constants
+    use crate::vga::{BUFFER_HEIGHT, BUFFER_WIDTH, Color, ColorCode, ScreenChar, TextBufferOperations};
 
     // Mock VgaBuffer that uses a Vec for testing
     struct MockVgaBuffer {
